+++ CHANGELOG.md
@@ -42,6 +42,7 @@
 ### Removed
 - Playlists: daily jams setting, since now it's always active in ListenBrainz. There is no need anymore to manually enable it.
 ### Fixed
+- API: updated with latest ListenBrainz API changes.
 - UI: tooltip cut to 50 chars per tag line.
 - Configuration: .json files at 'foobar2000\js_data\presets\global' not being saved with the calculated properties based on user values from other files.
 - Fixed possible crash handling web request while closing foobar2000. See [this](https://hydrogenaud.io/index.php/topic,121047.msg1044579.html#msg1044579), although current methods don't use 'WinHttp.WinHttpRequest.5.1' but 'Microsoft.XMLHTTP' which hasn't given any problems yet.